refactor(app): split connection.ts into domain modules
Extract the monolithic 2852-line connection store into focused modules:
- types.ts (416 lines): all shared interfaces and type definitions
- utils.ts (86 lines): pure utility functions (stripAnsi, filterThinking, nextMessageId)
- message-handler.ts (1570 lines): handleMessage dispatch, module-level state (delta buffers, replay flags, heartbeat, message queue), session update helpers
- connection.ts (980 lines): store definition and actions

Backward compatible — all existing imports from store/connection continue to work via re-exports. Zero changes to public API or runtime behavior.
