+++ packages/app/src/store/message-handler.ts
@@ -1,0 +1,1574 @@
+/**
+ * WebSocket message handler — processes all incoming server messages.
+ *
+ * Extracted from connection.ts to reduce file size. This module contains:
+ * - handleMessage() — the main message dispatch (~1100 lines)
+ * - Module-level state (delta buffers, replay flags, heartbeat, message queue)
+ * - Session update helpers (updateSession, updateActiveSession)
+ *
+ * Depends on the Zustand store via a late-bound reference (setStore) to
+ * avoid circular imports.
+ */
+import { Alert } from 'react-native';
+import {
+  createKeyPair,
+  deriveSharedKey,
+  DIRECTION_CLIENT,
+  DIRECTION_SERVER,
+  type EncryptionState,
+  type KeyPair,
+  type EncryptedEnvelope,
+} from '../utils/crypto';
+import { registerForPushNotifications } from '../notifications';
+import { stripAnsi, filterThinking, nextMessageId } from './utils';
+import type {
+  ChatMessage,
+  ConnectedClient,
+  ConnectionContext,
+  ConnectionState,
+  CustomAgent,
+  DiffFile,
+  DiscoveredSession,
+  DirectoryEntry,
+  FileEntry,
+  McpServer,
+  ModelInfo,
+  QueuedMessage,
+  ServerError,
+  SessionInfo,
+  SessionNotification,
+  SessionState,
+  SlashCommand,
+} from './types';
+import { createEmptySessionState } from './utils';
+
+// ---------------------------------------------------------------------------
+// Late-bound store reference — set once by connection.ts after store creation
+// ---------------------------------------------------------------------------
+type StoreApi = {
+  getState: () => ConnectionState;
+  setState: (s: Partial<ConnectionState> | ((state: ConnectionState) => Partial<ConnectionState>)) => void;
+};
+let _store: StoreApi | null = null;
+
+export function setStore(store: StoreApi): void {
+  _store = store;
+}
+
+function getStore(): StoreApi {
+  if (!_store) throw new Error('Store not initialized — call setStore() first');
+  return _store;
+}
+
+// ---------------------------------------------------------------------------
+// E2E encryption state — reset on every new connection
+// ---------------------------------------------------------------------------
+let _encryptionState: EncryptionState | null = null;
+let _pendingKeyPair: KeyPair | null = null;
+
+/**
+ * Send a JSON message over WebSocket, encrypting if E2E encryption is active.
+ * Use this instead of raw `socket.send(JSON.stringify(...))`.
+ */
+export function wsSend(socket: WebSocket, payload: Record<string, unknown>): void {
+  if (_encryptionState) {
+    const envelope = encrypt(JSON.stringify(payload), _encryptionState.sharedKey, _encryptionState.sendNonce, DIRECTION_CLIENT);
+    _encryptionState.sendNonce++;
+    socket.send(JSON.stringify(envelope));
+  } else {
+    socket.send(JSON.stringify(payload));
+  }
+}
+
+// Re-export encrypt for wsSend (import is used inside the function)
+import { encrypt, decrypt } from '../utils/crypto';
+
+// ---------------------------------------------------------------------------
+// Connection context (set by connect(), read by handleMessage)
+// ---------------------------------------------------------------------------
+let _connectionContext: ConnectionContext | null = null;
+
+export function setConnectionContext(ctx: ConnectionContext | null): void {
+  _connectionContext = ctx;
+}
+
+export function getConnectionContext(): ConnectionContext | null {
+  return _connectionContext;
+}
+
+// ---------------------------------------------------------------------------
+// Encryption state accessors
+// ---------------------------------------------------------------------------
+export function getEncryptionState(): EncryptionState | null {
+  return _encryptionState;
+}
+
+export function setEncryptionState(state: EncryptionState | null): void {
+  _encryptionState = state;
+}
+
+export function getPendingKeyPair(): KeyPair | null {
+  return _pendingKeyPair;
+}
+
+export function setPendingKeyPair(kp: KeyPair | null): void {
+  _pendingKeyPair = kp;
+}
+
+// ---------------------------------------------------------------------------
+// Connection attempt tracking
+// ---------------------------------------------------------------------------
+export let connectionAttemptId = 0;
+export let disconnectedAttemptId = -1;
+export let lastConnectedUrl: string | null = null;
+
+export function bumpConnectionAttemptId(): number {
+  return ++connectionAttemptId;
+}
+
+export function setDisconnectedAttemptId(id: number): void {
+  disconnectedAttemptId = id;
+}
+
+export function setLastConnectedUrl(url: string | null): void {
+  lastConnectedUrl = url;
+}
+
+// ---------------------------------------------------------------------------
+// History replay flags
+// ---------------------------------------------------------------------------
+let _receivingHistoryReplay = false;
+let _isSessionSwitchReplay = false;
+let _pendingSwitchSessionId: string | null = null;
+
+export function setPendingSwitchSessionId(id: string | null): void {
+  _pendingSwitchSessionId = id;
+}
+
+export function resetReplayFlags(): void {
+  _receivingHistoryReplay = false;
+  _isSessionSwitchReplay = false;
+  _pendingSwitchSessionId = null;
+}
+
+// ---------------------------------------------------------------------------
+// Permission boundary message splitting (#554)
+// ---------------------------------------------------------------------------
+const _postPermissionSplits = new Set<string>();
+const _deltaIdRemaps = new Map<string, string>();
+
+export function clearPermissionSplits(): void {
+  _postPermissionSplits.clear();
+  _deltaIdRemaps.clear();
+}
+
+// ---------------------------------------------------------------------------
+// Terminal write batching
+// ---------------------------------------------------------------------------
+let _pendingTerminalWrites = '';
+let _terminalWriteTimer: ReturnType<typeof setTimeout> | null = null;
+
+export function flushTerminalWrites(): void {
+  _terminalWriteTimer = null;
+  if (_pendingTerminalWrites.length === 0) return;
+  const data = _pendingTerminalWrites;
+  _pendingTerminalWrites = '';
+  const cb = getStore().getState()._terminalWriteCallback;
+  if (cb) cb(data);
+}
+
+export function appendPendingTerminalWrite(data: string): void {
+  _pendingTerminalWrites += data;
+  if (!_terminalWriteTimer) {
+    _terminalWriteTimer = setTimeout(flushTerminalWrites, 50);
+  }
+}
+
+export function clearTerminalWriteBatching(): void {
+  if (_terminalWriteTimer) {
+    clearTimeout(_terminalWriteTimer);
+    _terminalWriteTimer = null;
+  }
+  _pendingTerminalWrites = '';
+}
+
+// ---------------------------------------------------------------------------
+// Client-side heartbeat
+// ---------------------------------------------------------------------------
+let _heartbeatInterval: ReturnType<typeof setInterval> | null = null;
+let _pongTimeout: ReturnType<typeof setTimeout> | null = null;
+const HEARTBEAT_INTERVAL_MS = 15_000;
+const PONG_TIMEOUT_MS = 5_000;
+
+export function stopHeartbeat(): void {
+  if (_heartbeatInterval) { clearInterval(_heartbeatInterval); _heartbeatInterval = null; }
+  if (_pongTimeout) { clearTimeout(_pongTimeout); _pongTimeout = null; }
+}
+
+export function startHeartbeat(socket: WebSocket): void {
+  stopHeartbeat();
+  _heartbeatInterval = setInterval(() => {
+    if (socket.readyState !== WebSocket.OPEN) { stopHeartbeat(); return; }
+    try {
+      wsSend(socket, { type: 'ping' });
+    } catch { stopHeartbeat(); return; }
+    _pongTimeout = setTimeout(() => {
+      console.warn('[ws] Heartbeat pong timeout — closing dead connection');
+      stopHeartbeat();
+      try { socket.close(); } catch {}
+    }, PONG_TIMEOUT_MS);
+  }, HEARTBEAT_INTERVAL_MS);
+}
+
+function _onPong(): void {
+  if (_pongTimeout) { clearTimeout(_pongTimeout); _pongTimeout = null; }
+}
+
+// ---------------------------------------------------------------------------
+// Delta batching
+// ---------------------------------------------------------------------------
+const pendingDeltas = new Map<string, { sessionId: string | null; delta: string }>();
+let deltaFlushTimer: ReturnType<typeof setTimeout> | null = null;
+
+function flushPendingDeltas(): void {
+  deltaFlushTimer = null;
+  if (pendingDeltas.size === 0) return;
+  const updates = new Map(pendingDeltas);
+  pendingDeltas.clear();
+
+  const state = getStore().getState();
+
+  const bySession = new Map<string | null, Map<string, string>>();
+  for (const [msgId, { sessionId, delta }] of updates) {
+    if (!bySession.has(sessionId)) bySession.set(sessionId, new Map());
+    bySession.get(sessionId)!.set(msgId, delta);
+  }
+
+  let newSessionStates = { ...state.sessionStates };
+  let flatUpdated = false;
+
+  for (const [sessionId, deltas] of bySession) {
+    if (sessionId && newSessionStates[sessionId]) {
+      const sessionState = newSessionStates[sessionId];
+      const updatedMessages = sessionState.messages.map((m) => {
+        const d = deltas.get(m.id);
+        return d ? { ...m, content: m.content + d } : m;
+      });
+      newSessionStates = {
+        ...newSessionStates,
+        [sessionId]: { ...sessionState, messages: updatedMessages },
+      };
+      if (sessionId === state.activeSessionId) {
+        getStore().setState({ sessionStates: newSessionStates, messages: updatedMessages });
+        flatUpdated = true;
+      }
+    } else {
+      getStore().setState((s) => ({
+        messages: s.messages.map((m) => {
+          const d = deltas.get(m.id);
+          return d ? { ...m, content: m.content + d } : m;
+        }),
+      }));
+      flatUpdated = true;
+    }
+  }
+
+  if (!flatUpdated) {
+    getStore().setState({ sessionStates: newSessionStates });
+  }
+}
+
+export function clearDeltaBuffers(): void {
+  if (deltaFlushTimer) {
+    clearTimeout(deltaFlushTimer);
+    deltaFlushTimer = null;
+  }
+  pendingDeltas.clear();
+}
+
+// ---------------------------------------------------------------------------
+// Message queue: buffer messages while disconnected, drain on reconnect
+// ---------------------------------------------------------------------------
+const QUEUE_TTLS: Record<string, number> = {
+  input: 60_000,
+  interrupt: 5_000,
+  permission_response: 300_000,
+  user_question_response: 60_000,
+};
+const QUEUE_MAX_SIZE = 10;
+const QUEUE_EXCLUDED = new Set(['set_model', 'set_permission_mode', 'mode', 'resize']);
+const messageQueue: QueuedMessage[] = [];
+
+export function enqueueMessage(type: string, payload: unknown): 'queued' | false {
+  if (QUEUE_EXCLUDED.has(type)) return false;
+  const maxAge = QUEUE_TTLS[type];
+  if (!maxAge) return false;
+  if (messageQueue.length >= QUEUE_MAX_SIZE) return false;
+  messageQueue.push({ type, payload, queuedAt: Date.now(), maxAge });
+  console.log(`[queue] Queued ${type} (${messageQueue.length}/${QUEUE_MAX_SIZE})`);
+  return 'queued';
+}
+
+export function drainMessageQueue(socket: WebSocket): void {
+  if (messageQueue.length === 0) return;
+  const now = Date.now();
+  const valid = messageQueue.filter((m) => now - m.queuedAt < m.maxAge);
+  messageQueue.length = 0;
+  if (valid.length === 0) return;
+  console.log(`[queue] Draining ${valid.length} queued message(s)`);
+  for (const m of valid) {
+    try {
+      wsSend(socket, m.payload as Record<string, unknown>);
+    } catch (err) {
+      console.warn(`[queue] Failed to send queued ${m.type}:`, err);
+    }
+  }
+}
+
+export function clearMessageQueue(): void {
+  messageQueue.length = 0;
+}
+
+/** @internal Exposed for testing only */
+export const _testQueueInternals = {
+  getQueue: () => messageQueue,
+  enqueue: enqueueMessage,
+  drain: drainMessageQueue,
+  clear: () => { messageQueue.length = 0; },
+};
+
+// ---------------------------------------------------------------------------
+// Session update helpers
+// ---------------------------------------------------------------------------
+
+/**
+ * Update any session's state by ID. Syncs to flat state only when the target
+ * session is the currently active session (so UI reads remain correct).
+ */
+export function updateSession(sessionId: string, updater: (session: SessionState) => Partial<SessionState>): void {
+  const state = getStore().getState();
+  if (!state.sessionStates[sessionId]) return;
+
+  const current = state.sessionStates[sessionId];
+  const patch = updater(current);
+  if (Object.keys(patch).length === 0) return;
+  const updated = { ...current, ...patch };
+  const newSessionStates = { ...state.sessionStates, [sessionId]: updated };
+
+  if (sessionId === state.activeSessionId) {
+    const flatPatch: Record<string, unknown> = { sessionStates: newSessionStates };
+    if ('messages' in patch) flatPatch.messages = patch.messages;
+    if ('streamingMessageId' in patch) flatPatch.streamingMessageId = patch.streamingMessageId;
+    if ('claudeReady' in patch) flatPatch.claudeReady = patch.claudeReady;
+    if ('activeModel' in patch) flatPatch.activeModel = patch.activeModel;
+    if ('permissionMode' in patch) flatPatch.permissionMode = patch.permissionMode;
+    if ('contextUsage' in patch) flatPatch.contextUsage = patch.contextUsage;
+    if ('lastResultCost' in patch) flatPatch.lastResultCost = patch.lastResultCost;
+    if ('lastResultDuration' in patch) flatPatch.lastResultDuration = patch.lastResultDuration;
+    if ('isIdle' in patch) flatPatch.isIdle = patch.isIdle;
+    getStore().setState(flatPatch);
+  } else {
+    getStore().setState({ sessionStates: newSessionStates });
+  }
+}
+
+/** Helper to update the active session's state and sync to flat state */
+export function updateActiveSession(updater: (session: SessionState) => Partial<SessionState>): void {
+  const state = getStore().getState();
+  const activeId = state.activeSessionId;
+  if (activeId) updateSession(activeId, updater);
+}
+
+// ---------------------------------------------------------------------------
+// Session notification helper
+// ---------------------------------------------------------------------------
+
+/**
+ * Push a notification for a background session event.
+ * Deduplicates by (sessionId, eventType) — replaces existing rather than stacking.
+ */
+function pushSessionNotification(
+  sessionId: string,
+  eventType: SessionNotification['eventType'],
+  message: string,
+): void {
+  const state = getStore().getState();
+  if (sessionId === state.activeSessionId) return;
+  const sessionInfo = state.sessions.find((s) => s.sessionId === sessionId);
+  const sessionName = sessionInfo?.name || sessionId;
+  const notification: SessionNotification = {
+    id: `${sessionId}-${eventType}-${Date.now()}`,
+    sessionId,
+    sessionName,
+    eventType,
+    message,
+    timestamp: Date.now(),
+  };
+  getStore().setState((s) => {
+    const filtered = s.sessionNotifications.filter(
+      (n) => !(n.sessionId === sessionId && n.eventType === eventType),
+    );
+    return { sessionNotifications: [...filtered, notification] };
+  });
+}
+
+// ---------------------------------------------------------------------------
+// Push token registration
+// ---------------------------------------------------------------------------
+
+async function registerPushToken(socket: WebSocket): Promise<void> {
+  try {
+    const token = await registerForPushNotifications();
+    if (token && socket.readyState === WebSocket.OPEN) {
+      wsSend(socket, { type: 'register_push_token', token });
+      console.log('[push] Registered push token with server');
+    }
+  } catch (err) {
+    console.log('[push] Push registration skipped:', err);
+  }
+}
+
+// ---------------------------------------------------------------------------
+// Connection persistence helpers
+// ---------------------------------------------------------------------------
+import * as SecureStore from 'expo-secure-store';
+
+const STORAGE_KEY_URL = 'chroxy_last_url';
+const STORAGE_KEY_TOKEN = 'chroxy_last_token';
+
+export async function saveConnection(url: string, token: string): Promise<void> {
+  try {
+    await SecureStore.setItemAsync(STORAGE_KEY_URL, url);
+    await SecureStore.setItemAsync(STORAGE_KEY_TOKEN, token);
+  } catch {
+    // Storage not available (e.g. Expo Go limitations)
+  }
+}
+
+export async function loadConnection(): Promise<{ url: string; token: string } | null> {
+  try {
+    const url = await SecureStore.getItemAsync(STORAGE_KEY_URL);
+    const token = await SecureStore.getItemAsync(STORAGE_KEY_TOKEN);
+    if (url && token) return { url, token };
+  } catch {
+    // Storage not available
+  }
+  return null;
+}
+
+export async function clearConnection(): Promise<void> {
+  try {
+    await SecureStore.deleteItemAsync(STORAGE_KEY_URL);
+    await SecureStore.deleteItemAsync(STORAGE_KEY_TOKEN);
+  } catch {
+    // Storage not available
+  }
+}
+
+// ---------------------------------------------------------------------------
+// handleMessage — main message dispatch
+// ---------------------------------------------------------------------------
+
+/**
+ * Handles a parsed WebSocket message. Extracted from the socket.onmessage
+ * closure so it can be tested directly with raw JSON payloads.
+ *
+ * Reads/writes store via getStore().getState()/setState() and
+ * module-level helpers (updateSession, updateActiveSession, nextMessageId, etc).
+ * The few variables that were closured in connect() are accessed via _connectionContext.
+ */
+export function handleMessage(raw: unknown, ctxOverride?: ConnectionContext): void {
+  const ctx = ctxOverride ?? _connectionContext;
+  if (!ctx) return;
+  if (!raw || typeof raw !== 'object' || Array.isArray(raw)) return;
+  const msg = raw as Record<string, unknown>;
+  if (typeof msg.type !== 'string') return;
+
+  const get = () => getStore().getState();
+  const set: (s: Partial<ConnectionState> | ((state: ConnectionState) => Partial<ConnectionState>)) => void =
+    (s) => getStore().setState(s as ConnectionState);
+
+  switch (msg.type) {
+    case 'pong':
+      _onPong();
+      return;
+
+    case 'auth_ok': {
+      // Reset replay flags — fresh auth means clean slate
+      _receivingHistoryReplay = false;
+      _isSessionSwitchReplay = false;
+      _pendingSwitchSessionId = null;
+      // Track this URL as successfully connected
+      lastConnectedUrl = ctx.url;
+      // Extract server context from auth_ok
+      const authServerMode: 'cli' | 'terminal' | null =
+        msg.serverMode === 'cli' || msg.serverMode === 'terminal' ? msg.serverMode : null;
+      const authSessionCwd = typeof msg.cwd === 'string' ? msg.cwd : null;
+      const authServerVersion = typeof msg.serverVersion === 'string' ? msg.serverVersion : null;
+      const authLatestVersion = typeof msg.latestVersion === 'string' ? msg.latestVersion : null;
+      const authServerCommit = typeof msg.serverCommit === 'string' ? msg.serverCommit : null;
+      // Parse connected clients list with self-detection via clientId
+      const myClientId = typeof msg.clientId === 'string' ? msg.clientId : null;
+      const rawClients = Array.isArray(msg.connectedClients) ? msg.connectedClients : [];
+      const clients: ConnectedClient[] = rawClients
+        .filter((c: unknown): c is { clientId: string } => !!c && typeof c === 'object' && typeof (c as Record<string, unknown>).clientId === 'string')
+        .map((c: { clientId: string; deviceName?: string; deviceType?: string; platform?: string }) => ({
+          clientId: c.clientId,
+          deviceName: typeof c.deviceName === 'string' ? c.deviceName : null,
+          deviceType: (['phone', 'tablet', 'desktop', 'unknown'].includes(c.deviceType ?? '') ? c.deviceType : 'unknown') as ConnectedClient['deviceType'],
+          platform: typeof c.platform === 'string' ? c.platform : 'unknown',
+          isSelf: c.clientId === myClientId,
+        }));
+
+      // On reconnect, preserve messages and terminal buffer
+      const connectedState = {
+        connectionPhase: 'connected' as const,
+        wsUrl: ctx.url,
+        apiToken: ctx.token,
+        socket: ctx.socket,
+        claudeReady: false,
+        serverMode: authServerMode,
+        sessionCwd: authSessionCwd,
+        serverVersion: authServerVersion,
+        latestVersion: authLatestVersion,
+        serverCommit: authServerCommit,
+        streamingMessageId: null,
+        myClientId: myClientId,
+        connectedClients: clients,
+        connectionError: null as string | null,
+        connectionRetryCount: 0,
+        // Clear shutdown state on successful connect
+        shutdownReason: null,
+        restartEtaMs: null,
+        restartingSince: null,
+      };
+      if (ctx.isReconnect) {
+        set(connectedState);
+      } else {
+        set({
+          ...connectedState,
+          messages: [],
+          terminalBuffer: '',
+          terminalRawBuffer: '',
+          sessions: [],
+          activeSessionId: null,
+          sessionStates: {},
+          customAgents: [],
+        });
+      }
+      // Start client-side heartbeat for dead connection detection
+      startHeartbeat(ctx.socket);
+
+      // Initiate key exchange if server requires encryption
+      if (msg.encryption === 'required') {
+        _pendingKeyPair = createKeyPair();
+        // Send key_exchange plaintext (before encryption is active)
+        ctx.socket.send(JSON.stringify({ type: 'key_exchange', publicKey: _pendingKeyPair.publicKey }));
+        // Post-auth messages will be sent after key_exchange_ok arrives
+      } else {
+        // No encryption — send post-auth messages immediately
+        wsSend(ctx.socket, { type: 'mode', mode: get().viewMode });
+        wsSend(ctx.socket, { type: 'list_slash_commands' });
+        wsSend(ctx.socket, { type: 'list_agents' });
+      }
+      // Save for quick reconnect
+      saveConnection(ctx.url, ctx.token);
+      set({ savedConnection: { url: ctx.url, token: ctx.token } });
+      // Register push token (async, non-blocking)
+      void registerPushToken(ctx.socket);
+      break;
+    }
+
+    case 'key_exchange_ok': {
+      if (_pendingKeyPair) {
+        if (!msg.publicKey || typeof msg.publicKey !== 'string') {
+          console.error('[crypto] Invalid publicKey in key_exchange_ok message', msg.publicKey);
+          ctx.socket.close();
+          set({ connectionPhase: 'disconnected', socket: null });
+          _pendingKeyPair = null;
+          break;
+        }
+        const sharedKey = deriveSharedKey(msg.publicKey, _pendingKeyPair.secretKey);
+        _encryptionState = { sharedKey, sendNonce: 0, recvNonce: 0 };
+        _pendingKeyPair = null;
+        console.log('[crypto] E2E encryption established');
+        // Now send the post-auth messages that were deferred
+        wsSend(ctx.socket, { type: 'mode', mode: get().viewMode });
+        wsSend(ctx.socket, { type: 'list_slash_commands' });
+        wsSend(ctx.socket, { type: 'list_agents' });
+      }
+      break;
+    }
+
+    case 'auth_fail':
+      ctx.socket.close();
+      set({ connectionPhase: 'disconnected', socket: null });
+      if (!ctx.silent) {
+        Alert.alert('Auth Failed', (msg.reason as string) || 'Invalid token');
+      }
+      break;
+
+    case 'server_mode':
+      set({ serverMode: msg.mode as 'cli' | 'terminal' });
+      // Force chat view in CLI mode (no terminal available)
+      if (msg.mode === 'cli' && get().viewMode === 'terminal') {
+        set({ viewMode: 'chat' });
+      }
+      break;
+
+    // --- Multi-session messages ---
+
+    case 'session_list':
+      if (Array.isArray(msg.sessions)) {
+        const sessionList = msg.sessions as SessionInfo[];
+        set({ sessions: sessionList });
+        // Sync conversationId from session list into session states
+        for (const s of sessionList) {
+          if (s.conversationId && get().sessionStates[s.sessionId]) {
+            updateSession(s.sessionId, (ss) =>
+              ss.conversationId !== s.conversationId ? { conversationId: s.conversationId } : {}
+            );
+          }
+        }
+      }
+      break;
+
+    case 'session_context': {
+      const ctxSessionId = (msg.sessionId as string) || get().activeSessionId;
+      if (ctxSessionId && get().sessionStates[ctxSessionId]) {
+        updateSession(ctxSessionId, () => ({
+          sessionContext: {
+            gitBranch: typeof msg.gitBranch === 'string' ? msg.gitBranch : null,
+            gitDirty: typeof msg.gitDirty === 'number' ? msg.gitDirty : 0,
+            gitAhead: typeof msg.gitAhead === 'number' ? msg.gitAhead : 0,
+            projectName: typeof msg.projectName === 'string' ? msg.projectName : null,
+          },
+        }));
+      }
+      break;
+    }
+
+    case 'session_switched': {
+      const sessionId = msg.sessionId as string;
+      // Only treat as session-switch replay if the user explicitly initiated it
+      // (auth-triggered session_switched on reconnect should use reconnect dedup)
+      if (_pendingSwitchSessionId && _pendingSwitchSessionId === sessionId) {
+        _isSessionSwitchReplay = true;
+      }
+      _pendingSwitchSessionId = null;
+      const switchConvId = typeof msg.conversationId === 'string' ? msg.conversationId : null;
+      set((state: ConnectionState) => {
+        // Initialize session state if it doesn't exist
+        const sessionStates = { ...state.sessionStates };
+        if (!sessionStates[sessionId]) {
+          sessionStates[sessionId] = createEmptySessionState();
+        }
+        // Update conversationId if provided
+        if (switchConvId) {
+          sessionStates[sessionId] = { ...sessionStates[sessionId], conversationId: switchConvId };
+        }
+        const ss = sessionStates[sessionId];
+        return {
+          activeSessionId: sessionId,
+          sessionStates,
+          // Sync flat state from the switched-to session
+          messages: ss.messages,
+          streamingMessageId: ss.streamingMessageId,
+          claudeReady: ss.claudeReady,
+          activeModel: ss.activeModel,
+          permissionMode: ss.permissionMode,
+          contextUsage: ss.contextUsage,
+          lastResultCost: ss.lastResultCost,
+          lastResultDuration: ss.lastResultDuration,
+          isIdle: ss.isIdle,
+        };
+      });
+      // Refresh slash commands (project commands may differ per session cwd)
+      get().fetchSlashCommands();
+      // Refresh agents (project agents may differ per session cwd)
+      get().fetchCustomAgents();
+      break;
+    }
+
+    case 'conversation_id': {
+      const convSessionId = msg.sessionId as string;
+      const conversationId = typeof msg.conversationId === 'string' ? msg.conversationId : null;
+      if (convSessionId && get().sessionStates[convSessionId]) {
+        updateSession(convSessionId, () => ({ conversationId }));
+      }
+      break;
+    }
+
+    case 'session_error': {
+      const errorSessionId = (msg.sessionId as string) || get().activeSessionId;
+      if (msg.category === 'crash' && errorSessionId && get().sessionStates[errorSessionId]) {
+        updateSession(errorSessionId, () => ({ health: 'crashed' as const }));
+        pushSessionNotification(errorSessionId, 'error', 'Session crashed');
+      }
+      if (msg.category !== 'crash') {
+        Alert.alert('Session Error', (msg.message as string) || 'Unknown error');
+      }
+      break;
+    }
+
+    case 'discovered_sessions':
+      if (Array.isArray(msg.tmux)) {
+        set({ discoveredSessions: msg.tmux as DiscoveredSession[] });
+        if ((msg.tmux as DiscoveredSession[]).length > 0) {
+          const names = (msg.tmux as DiscoveredSession[]).map((s: DiscoveredSession) => s.sessionName).join(', ');
+          const discoveryMsg: ChatMessage = {
+            id: nextMessageId('discovery'),
+            type: 'system',
+            content: (msg.tmux as DiscoveredSession[]).length === 1
+              ? `New Claude session found: ${names}. Open session picker to attach.`
+              : `${(msg.tmux as DiscoveredSession[]).length} new Claude sessions found: ${names}. Open session picker to attach.`,
+            timestamp: Date.now(),
+          };
+          const activeId = get().activeSessionId;
+          if (activeId && get().sessionStates[activeId]) {
+            updateActiveSession((ss) => ({
+              messages: [...ss.messages, discoveryMsg],
+            }));
+          } else {
+            get().addMessage(discoveryMsg);
+          }
+        }
+      }
+      break;
+
+    // --- History replay ---
+
+    case 'history_replay_start':
+      _receivingHistoryReplay = true;
+      // Full history replay (from request_full_history): clear messages before replay
+      if (msg.fullHistory === true) {
+        _isSessionSwitchReplay = true;
+        const targetId = (msg.sessionId as string) || get().activeSessionId;
+        if (targetId && get().sessionStates[targetId]) {
+          updateSession(targetId, () => ({ messages: [] }));
+        }
+      }
+      // Clear transient state — these events are not replayed from history,
+      // so any surviving entries are stale from pre-disconnect
+      updateActiveSession((ss) => {
+        const patch: Partial<SessionState> = {};
+        if (ss.activeAgents.length > 0) patch.activeAgents = [];
+        if (ss.isPlanPending) {
+          patch.isPlanPending = false;
+          patch.planAllowedPrompts = [];
+        }
+        return Object.keys(patch).length > 0 ? patch : {};
+      });
+      break;
+
+    case 'history_replay_end':
+      _receivingHistoryReplay = false;
+      _isSessionSwitchReplay = false;
+      // Mark all replayed prompts as answered — any prompt in history
+      // has already been resolved by the server.
+      updateActiveSession((ss) => {
+        const hasUnansweredPrompts = ss.messages.some(
+          (m) => m.type === 'prompt' && !m.answered
+        );
+        if (!hasUnansweredPrompts) return {};
+        return {
+          messages: ss.messages.map((m) =>
+            m.type === 'prompt' && !m.answered
+              ? { ...m, answered: '(resolved)' }
+              : m
+          ),
+        };
+      });
+      break;
+
+    // --- Existing message handlers (now session-aware) ---
+
+    case 'message': {
+      const msgType = (msg.messageType || msg.type) as string;
+      // Skip server-echoed user_input — we already show it instantly client-side
+      // But allow user_input during full history sync (messages came from terminal)
+      if (msgType === 'user_input' && !(_receivingHistoryReplay && _isSessionSwitchReplay)) break;
+      const targetId = (msg.sessionId as string) || get().activeSessionId;
+      // During reconnect replay, skip if app already has messages (cache is fresh)
+      if (_receivingHistoryReplay && !_isSessionSwitchReplay && get().messages.length > 0) break;
+      // During session-switch replay, skip if an equivalent message is already in cache (dedup)
+      if (_receivingHistoryReplay && _isSessionSwitchReplay) {
+        const targetState = targetId ? get().sessionStates[targetId] : null;
+        const cached = targetState ? targetState.messages : get().messages;
+        const isDuplicate = cached.some((m) => {
+          if (m.type !== msgType || m.content !== msg.content) return false;
+          if (m.timestamp !== msg.timestamp) return false;
+          if ((m.tool ?? null) !== (msg.tool ?? null)) return false;
+          return JSON.stringify(m.options ?? null) === JSON.stringify(msg.options ?? null);
+        });
+        if (isDuplicate) break;
+      }
+      const newMsg: ChatMessage = {
+        id: nextMessageId(msgType),
+        type: msgType as ChatMessage['type'],
+        content: msg.content as string,
+        tool: msg.tool as string | undefined,
+        options: msg.options as ChatMessage['options'],
+        timestamp: msg.timestamp as number,
+      };
+      if (targetId && get().sessionStates[targetId]) {
+        updateSession(targetId, (ss) => ({
+          messages: [
+            ...ss.messages.filter((m) => m.id !== 'thinking' || newMsg.id === 'thinking'),
+            newMsg,
+          ],
+        }));
+      } else {
+        get().addMessage(newMsg);
+      }
+      break;
+    }
+
+    case 'stream_start': {
+      const streamId = msg.messageId as string;
+      const targetId = (msg.sessionId as string) || get().activeSessionId;
+      if (targetId && get().sessionStates[targetId]) {
+        updateSession(targetId, (ss) => {
+          if (ss.messages.some((m) => m.id === streamId)) {
+            return { streamingMessageId: streamId };
+          }
+          return {
+            streamingMessageId: streamId,
+            messages: [
+              ...filterThinking(ss.messages),
+              { id: streamId, type: 'response' as const, content: '', timestamp: Date.now() },
+            ],
+          };
+        });
+      } else {
+        set((state: ConnectionState) => {
+          if (state.messages.some((m) => m.id === streamId)) {
+            return { streamingMessageId: streamId };
+          }
+          return {
+            streamingMessageId: streamId,
+            messages: [
+              ...filterThinking(state.messages),
+              { id: streamId, type: 'response' as const, content: '', timestamp: Date.now() },
+            ],
+          };
+        });
+      }
+      break;
+    }
+
+    case 'stream_delta': {
+      let deltaId = msg.messageId as string;
+      const capturedSessionId = (msg.sessionId as string) || get().activeSessionId;
+
+      // Permission boundary split: first delta after a split creates a new message
+      if (_postPermissionSplits.has(deltaId)) {
+        _postPermissionSplits.delete(deltaId);
+        const newId = `${deltaId}-post-${Date.now()}`;
+        _deltaIdRemaps.set(deltaId, newId);
+        const newMsg: ChatMessage = {
+          id: newId,
+          type: 'response',
+          content: '',
+          timestamp: Date.now(),
+        };
+        const targetId = capturedSessionId;
+        if (targetId && get().sessionStates[targetId]) {
+          updateSession(targetId, (ss) => ({
+            streamingMessageId: newId,
+            messages: [...ss.messages, newMsg],
+          }));
+        } else {
+          set((state: ConnectionState) => ({
+            streamingMessageId: newId,
+            messages: [...state.messages, newMsg],
+          }));
+        }
+        deltaId = newId;
+      } else if (_deltaIdRemaps.has(deltaId)) {
+        deltaId = _deltaIdRemaps.get(deltaId)!;
+      }
+
+      const existingDelta = pendingDeltas.get(deltaId);
+      pendingDeltas.set(deltaId, {
+        sessionId: capturedSessionId,
+        delta: (existingDelta?.delta || '') + (msg.delta as string),
+      });
+      if (!deltaFlushTimer) {
+        deltaFlushTimer = setTimeout(flushPendingDeltas, 100);
+      }
+      break;
+    }
+
+    case 'stream_end':
+      // Flush any buffered deltas immediately before clearing streaming state
+      if (deltaFlushTimer) {
+        clearTimeout(deltaFlushTimer);
+      }
+      flushPendingDeltas();
+      // Clean up permission boundary split tracking
+      _postPermissionSplits.delete(msg.messageId as string);
+      _deltaIdRemaps.delete(msg.messageId as string);
+      {
+        const targetId = (msg.sessionId as string) || get().activeSessionId;
+        if (targetId && get().sessionStates[targetId]) {
+          updateSession(targetId, () => ({ streamingMessageId: null }));
+        } else {
+          set({ streamingMessageId: null });
+        }
+      }
+      break;
+
+    case 'tool_start': {
+      const targetId = (msg.sessionId as string) || get().activeSessionId;
+      // During reconnect replay, skip if app already has messages (cache is fresh)
+      if (_receivingHistoryReplay && !_isSessionSwitchReplay && get().messages.length > 0) break;
+      // Use server messageId as stable identifier for dedup (same ID on live + replay)
+      const toolId = (msg.messageId as string) || nextMessageId('tool');
+      // During session-switch replay, skip if tool already in cache (dedup by stable ID)
+      if (_receivingHistoryReplay && _isSessionSwitchReplay) {
+        const targetState = targetId ? get().sessionStates[targetId] : null;
+        const cached = targetState ? targetState.messages : get().messages;
+        if (cached.some((m) => m.id === toolId)) break;
+      }
+      const toolMsg: ChatMessage = {
+        id: toolId,
+        type: 'tool_use',
+        content: msg.input ? JSON.stringify(msg.input) : (msg.tool as string) || '',
+        tool: msg.tool as string | undefined,
+        toolUseId: msg.toolUseId as string | undefined,
+        serverName: msg.serverName as string | undefined,
+        timestamp: Date.now(),
+      };
+      if (targetId && get().sessionStates[targetId]) {
+        updateSession(targetId, (ss) => ({
+          messages: [...ss.messages, toolMsg],
+        }));
+      } else {
+        get().addMessage(toolMsg);
+      }
+      break;
+    }
+
+    case 'tool_result': {
+      const toolUseId = msg.toolUseId as string;
+      if (!toolUseId) break;
+      const resultText = (msg.result as string) || '';
+      const truncated = !!(msg.truncated as boolean);
+      const images = Array.isArray(msg.images) ? msg.images as { mediaType: string; data: string }[] : undefined;
+      const targetId = (msg.sessionId as string) || get().activeSessionId;
+      // Find the matching tool_use message and attach the result
+      const patch: Partial<ChatMessage> = { toolResult: resultText, toolResultTruncated: truncated };
+      if (images?.length) patch.toolResultImages = images;
+      const patchResult = (ss: SessionState) => {
+        const idx = ss.messages.findIndex(
+          (m) => m.type === 'tool_use' && m.toolUseId === toolUseId,
+        );
+        if (idx === -1) return {};
+        const updated = [...ss.messages];
+        updated[idx] = { ...updated[idx], ...patch };
+        return { messages: updated };
+      };
+      if (targetId && get().sessionStates[targetId]) {
+        updateSession(targetId, patchResult);
+      } else {
+        const idx = get().messages.findIndex(
+          (m) => m.type === 'tool_use' && m.toolUseId === toolUseId,
+        );
+        if (idx !== -1) {
+          const updated = [...get().messages];
+          updated[idx] = { ...updated[idx], ...patch };
+          set({ messages: updated });
+        }
+      }
+      break;
+    }
+
+    case 'result': {
+      // Flush any buffered deltas before clearing streaming state
+      if (deltaFlushTimer) {
+        clearTimeout(deltaFlushTimer);
+      }
+      flushPendingDeltas();
+      // Clean up permission boundary split tracking
+      _postPermissionSplits.clear();
+      _deltaIdRemaps.clear();
+      const usage = msg.usage as Record<string, number> | undefined;
+      const resultPatch = {
+        streamingMessageId: null as string | null,
+        contextUsage: usage
+          ? {
+              inputTokens: usage.input_tokens || 0,
+              outputTokens: usage.output_tokens || 0,
+              cacheCreation: usage.cache_creation_input_tokens || 0,
+              cacheRead: usage.cache_read_input_tokens || 0,
+            }
+          : null,
+        lastResultCost: typeof msg.cost === 'number' ? msg.cost : null,
+        lastResultDuration: typeof msg.duration === 'number' ? msg.duration : null,
+      };
+      const targetId = (msg.sessionId as string) || get().activeSessionId;
+      // Notify if a background session just finished (was streaming)
+      if (targetId && get().sessionStates[targetId]?.streamingMessageId) {
+        pushSessionNotification(targetId, 'completed', 'Task completed');
+      }
+      if (targetId && get().sessionStates[targetId]) {
+        updateSession(targetId, () => resultPatch);
+      } else {
+        set(resultPatch);
+      }
+      break;
+    }
+
+    case 'model_changed': {
+      const model = (typeof msg.model === 'string' && (msg.model as string).trim()) ? (msg.model as string).trim() : null;
+      const targetId = (msg.sessionId as string) || get().activeSessionId;
+      if (targetId && get().sessionStates[targetId]) {
+        updateSession(targetId, () => ({ activeModel: model }));
+      } else {
+        set({ activeModel: model });
+      }
+      break;
+    }
+
+    case 'available_models':
+      if (Array.isArray(msg.models)) {
+        const cleaned = (msg.models as unknown[])
+          .map((m: unknown): ModelInfo | null => {
+            if (typeof m === 'object' && m !== null) {
+              const { id, label, fullId } = m as ModelInfo;
+              if (
+                typeof id === 'string' && id.trim() !== '' &&
+                typeof label === 'string' && label.trim() !== '' &&
+                typeof fullId === 'string' && fullId.trim() !== ''
+              ) {
+                return { id, label, fullId };
+              }
+            }
+            if (typeof m === 'string' && m.trim().length > 0) {
+              const s = m.trim();
+              return { id: s, label: s.charAt(0).toUpperCase() + s.slice(1), fullId: s };
+            }
+            return null;
+          })
+          .filter((m: ModelInfo | null): m is ModelInfo => m !== null);
+        set({ availableModels: cleaned });
+      }
+      break;
+
+    case 'permission_mode_changed': {
+      const mode = (typeof msg.mode === 'string' && (msg.mode as string).trim()) ? (msg.mode as string).trim() : null;
+      const targetId = (msg.sessionId as string) || get().activeSessionId;
+      if (targetId && get().sessionStates[targetId]) {
+        updateSession(targetId, () => ({ permissionMode: mode }));
+      } else {
+        set({ permissionMode: mode });
+      }
+      // Clear pending confirm if mode change arrived (confirmation was accepted)
+      set({ pendingPermissionConfirm: null });
+      break;
+    }
+
+    case 'confirm_permission_mode': {
+      const confirmMode = typeof msg.mode === 'string' ? msg.mode : null;
+      const warning = typeof msg.warning === 'string' ? msg.warning : 'Are you sure?';
+      if (confirmMode) {
+        set({ pendingPermissionConfirm: { mode: confirmMode, warning } });
+      }
+      break;
+    }
+
+    case 'available_permission_modes':
+      if (Array.isArray(msg.modes)) {
+        const cleaned = (msg.modes as unknown[])
+          .filter((m): m is { id: string; label: string } =>
+            typeof m === 'object' && m !== null &&
+            typeof (m as { id: unknown }).id === 'string' &&
+            typeof (m as { label: unknown }).label === 'string'
+          );
+        set({ availablePermissionModes: cleaned });
+      }
+      break;
+
+    case 'status_update': {
+      const statusSid = (msg.sessionId as string) || get().activeSessionId;
+      if (statusSid && statusSid !== get().activeSessionId) break;
+      set({
+        claudeStatus: {
+          cost: msg.cost as number,
+          model: msg.model as string,
+          messageCount: msg.messageCount as number,
+          contextTokens: msg.contextTokens as string,
+          contextPercent: msg.contextPercent as number,
+          compactPercent: (msg.compactPercent as number) ?? null,
+        },
+      });
+      break;
+    }
+
+    case 'raw':
+      get().appendTerminalData(msg.data as string);
+      break;
+
+    case 'claude_ready': {
+      const targetId = (msg.sessionId as string) || get().activeSessionId;
+      if (targetId && get().sessionStates[targetId]) {
+        updateSession(targetId, () => ({ claudeReady: true }));
+      } else {
+        set({ claudeReady: true });
+      }
+      // Drain queued messages on reconnect
+      const readySocket = get().socket;
+      if (readySocket && readySocket.readyState === WebSocket.OPEN) {
+        drainMessageQueue(readySocket);
+      }
+      break;
+    }
+
+    case 'agent_idle': {
+      const idleTargetId = (msg.sessionId as string) || get().activeSessionId;
+      if (idleTargetId && get().sessionStates[idleTargetId]) {
+        updateSession(idleTargetId, () => ({ isIdle: true }));
+      }
+      break;
+    }
+
+    case 'agent_busy': {
+      const busyTargetId = (msg.sessionId as string) || get().activeSessionId;
+      if (busyTargetId && get().sessionStates[busyTargetId]) {
+        updateSession(busyTargetId, () => ({ isIdle: false }));
+      }
+      break;
+    }
+
+    case 'agent_spawned': {
+      const spawnTargetId = (msg.sessionId as string) || get().activeSessionId;
+      if (spawnTargetId && get().sessionStates[spawnTargetId]) {
+        updateSession(spawnTargetId, (ss) => {
+          if (ss.activeAgents.some((a) => a.toolUseId === msg.toolUseId)) return {};
+          return {
+            activeAgents: [...ss.activeAgents, {
+              toolUseId: msg.toolUseId as string,
+              description: (msg.description as string) || 'Background task',
+              startedAt: (msg.startedAt as number) || Date.now(),
+            }],
+          };
+        });
+      }
+      break;
+    }
+
+    case 'agent_completed': {
+      const completeTargetId = (msg.sessionId as string) || get().activeSessionId;
+      if (completeTargetId && get().sessionStates[completeTargetId]) {
+        updateSession(completeTargetId, (ss) => {
+          const filtered = ss.activeAgents.filter(
+            (a) => a.toolUseId !== msg.toolUseId
+          );
+          if (filtered.length === ss.activeAgents.length) return {};
+          return { activeAgents: filtered };
+        });
+      }
+      break;
+    }
+
+    case 'plan_started': {
+      const planStartTargetId = (msg.sessionId as string) || get().activeSessionId;
+      if (planStartTargetId && get().sessionStates[planStartTargetId]) {
+        updateSession(planStartTargetId, () => ({
+          isPlanPending: false,
+          planAllowedPrompts: [],
+        }));
+      }
+      break;
+    }
+
+    case 'plan_ready': {
+      const planReadyTargetId = (msg.sessionId as string) || get().activeSessionId;
+      const prompts = Array.isArray(msg.allowedPrompts) ? msg.allowedPrompts as { tool: string; prompt: string }[] : [];
+      if (planReadyTargetId && get().sessionStates[planReadyTargetId]) {
+        updateSession(planReadyTargetId, () => ({
+          isPlanPending: true,
+          planAllowedPrompts: prompts,
+        }));
+      }
+      break;
+    }
+
+    case 'raw_background':
+      get().appendTerminalData(msg.data as string);
+      break;
+
+    case 'permission_request': {
+      // Split streaming response at permission boundary (#554)
+      {
+        const permTargetId = (msg.sessionId as string) || get().activeSessionId;
+        const currentStreamId = permTargetId && get().sessionStates[permTargetId]
+          ? get().sessionStates[permTargetId].streamingMessageId
+          : get().streamingMessageId;
+        if (currentStreamId && currentStreamId !== 'pending') {
+          if (deltaFlushTimer) {
+            clearTimeout(deltaFlushTimer);
+          }
+          flushPendingDeltas();
+          let serverStreamId = currentStreamId;
+          for (const [origId, remappedId] of _deltaIdRemaps) {
+            if (remappedId === currentStreamId) {
+              serverStreamId = origId;
+              break;
+            }
+          }
+          _postPermissionSplits.add(serverStreamId);
+          if (permTargetId && get().sessionStates[permTargetId]) {
+            updateSession(permTargetId, () => ({ streamingMessageId: null }));
+          } else {
+            set({ streamingMessageId: null });
+          }
+        }
+      }
+      const permRequestId = msg.requestId as string;
+      const newOptions = [
+        { label: 'Allow', value: 'allow' },
+        { label: 'Deny', value: 'deny' },
+        { label: 'Always Allow', value: 'allowAlways' },
+      ];
+      const newExpiresAt = typeof msg.remainingMs === 'number' ? Date.now() + msg.remainingMs : undefined;
+      const permTargetId = (msg.sessionId as string) || get().activeSessionId;
+
+      const targetMessages = permTargetId && get().sessionStates[permTargetId]
+        ? get().sessionStates[permTargetId].messages
+        : get().messages;
+      const existingIdx = targetMessages.findIndex(
+        (m) => m.requestId === permRequestId && m.type === 'prompt'
+      );
+
+      if (existingIdx !== -1) {
+        const updater = (ss: { messages: ChatMessage[] }) => ({
+          messages: ss.messages.map((m) =>
+            m.requestId === permRequestId && m.type === 'prompt'
+              ? { ...m, answered: undefined, options: newOptions, expiresAt: newExpiresAt }
+              : m
+          ),
+        });
+        if (permTargetId && get().sessionStates[permTargetId]) {
+          updateSession(permTargetId, updater);
+        } else {
+          set({ messages: updater({ messages: get().messages }).messages });
+        }
+      } else {
+        const permMsg: ChatMessage = {
+          id: nextMessageId('perm'),
+          type: 'prompt',
+          content: msg.tool ? `${msg.tool}: ${msg.description}` : ((msg.description as string) || 'Permission required'),
+          tool: msg.tool as string | undefined,
+          requestId: permRequestId,
+          toolInput: msg.input && typeof msg.input === 'object' ? msg.input as Record<string, unknown> : undefined,
+          options: newOptions,
+          expiresAt: newExpiresAt,
+          timestamp: Date.now(),
+        };
+        if (permTargetId && get().sessionStates[permTargetId]) {
+          updateSession(permTargetId, (ss) => ({
+            messages: [...ss.messages, permMsg],
+          }));
+        } else {
+          get().addMessage(permMsg);
+        }
+      }
+      if (permTargetId) {
+        const toolDesc = msg.tool ? `${msg.tool}` : 'Permission needed';
+        pushSessionNotification(permTargetId, 'permission', toolDesc);
+      }
+      break;
+    }
+
+    case 'permission_expired': {
+      const expiredRequestId = msg.requestId as string;
+      if (expiredRequestId) {
+        console.warn(`[ws] Permission ${expiredRequestId} expired: ${msg.message}`);
+        const expTargetId = (msg.sessionId as string) || get().activeSessionId;
+        if (expTargetId && get().sessionStates[expTargetId]) {
+          updateSession(expTargetId, (ss) => ({
+            messages: ss.messages.map((m) =>
+              m.requestId === expiredRequestId && m.type === 'prompt'
+                ? { ...m, content: `${m.content}\n(Expired — this permission was already handled or timed out)`, options: undefined }
+                : m
+            ),
+          }));
+        }
+      }
+      break;
+    }
+
+    case 'user_question': {
+      const questions = msg.questions as unknown[];
+      if (!Array.isArray(questions) || questions.length === 0) break;
+      const q = questions[0] as Record<string, unknown>;
+      if (!q || typeof q !== 'object' || typeof q.question !== 'string') break;
+      const questionMsg: ChatMessage = {
+        id: nextMessageId('question'),
+        type: 'prompt',
+        content: q.question as string,
+        toolUseId: msg.toolUseId as string,
+        options: Array.isArray(q.options)
+          ? (q.options as unknown[])
+              .filter((o: unknown): o is { label: string } => !!o && typeof o === 'object' && typeof (o as Record<string, unknown>).label === 'string')
+              .map((o: { label: string }) => ({
+                label: o.label,
+                value: o.label,
+              }))
+          : [],
+        timestamp: Date.now(),
+      };
+      const questionTargetId = (msg.sessionId as string) || get().activeSessionId;
+      if (questionTargetId && get().sessionStates[questionTargetId]) {
+        updateSession(questionTargetId, (ss) => ({
+          messages: [...ss.messages, questionMsg],
+        }));
+      } else {
+        get().addMessage(questionMsg);
+      }
+      if (questionTargetId) {
+        const questionText = (q.question as string).slice(0, 60);
+        pushSessionNotification(questionTargetId, 'question', questionText);
+      }
+      break;
+    }
+
+    case 'server_status': {
+      const statusMessage: string =
+        typeof msg.message === 'string' && (msg.message as string).trim().length > 0
+          ? stripAnsi(msg.message as string)
+          : 'Status update';
+      const statusMsg: ChatMessage = {
+        id: nextMessageId('status'),
+        type: 'system',
+        content: statusMessage,
+        timestamp: Date.now(),
+      };
+      const activeStatusId = get().activeSessionId;
+      if (activeStatusId && get().sessionStates[activeStatusId]) {
+        updateActiveSession((ss) => ({
+          messages: [...ss.messages, statusMsg],
+        }));
+      } else {
+        get().addMessage(statusMsg);
+      }
+      break;
+    }
+
+    case 'server_shutdown': {
+      const reason = msg.reason === 'restart' || msg.reason === 'shutdown' ? msg.reason : 'shutdown';
+      const eta = typeof msg.restartEtaMs === 'number' ? msg.restartEtaMs : 0;
+      set({
+        shutdownReason: reason,
+        restartEtaMs: eta,
+        restartingSince: Date.now(),
+      });
+      break;
+    }
+
+    // --- Multi-client awareness ---
+
+    case 'client_joined': {
+      if (!msg.client || typeof (msg.client as Record<string, unknown>).clientId !== 'string') break;
+      const client = msg.client as Record<string, unknown>;
+      const newClient: ConnectedClient = {
+        clientId: client.clientId as string,
+        deviceName: typeof client.deviceName === 'string' ? client.deviceName : null,
+        deviceType: (['phone', 'tablet', 'desktop', 'unknown'].includes(client.deviceType as string) ? client.deviceType : 'unknown') as ConnectedClient['deviceType'],
+        platform: typeof client.platform === 'string' ? client.platform : 'unknown',
+        isSelf: false,
+      };
+      set((state: ConnectionState) => ({
+        connectedClients: [...state.connectedClients.filter((c) => c.clientId !== newClient.clientId), newClient],
+      }));
+      const deviceLabel = newClient.deviceName || 'A device';
+      const joinMsg: ChatMessage = {
+        id: nextMessageId('client'),
+        type: 'system',
+        content: `${deviceLabel} connected`,
+        timestamp: Date.now(),
+      };
+      const joinActiveId = get().activeSessionId;
+      if (joinActiveId && get().sessionStates[joinActiveId]) {
+        updateActiveSession((ss) => ({
+          messages: [...ss.messages, joinMsg],
+        }));
+      } else {
+        get().addMessage(joinMsg);
+      }
+      break;
+    }
+
+    case 'client_left': {
+      if (typeof msg.clientId !== 'string') break;
+      const departingClient = get().connectedClients.find((c) => c.clientId === msg.clientId);
+      set((state: ConnectionState) => ({
+        connectedClients: state.connectedClients.filter((c) => c.clientId !== msg.clientId),
+      }));
+      const leftLabel = departingClient?.deviceName || 'A device';
+      const leftMsg: ChatMessage = {
+        id: nextMessageId('client'),
+        type: 'system',
+        content: `${leftLabel} disconnected`,
+        timestamp: Date.now(),
+      };
+      const leftActiveId = get().activeSessionId;
+      if (leftActiveId && get().sessionStates[leftActiveId]) {
+        updateActiveSession((ss) => ({
+          messages: [...ss.messages, leftMsg],
+        }));
+      } else {
+        get().addMessage(leftMsg);
+      }
+      break;
+    }
+
+    case 'primary_changed': {
+      const primarySessionId = msg.sessionId as string;
+      const primaryClientId = typeof msg.clientId === 'string' ? msg.clientId : null;
+      if (typeof primarySessionId === 'string' && get().sessionStates[primarySessionId]) {
+        updateSession(primarySessionId, () => ({
+          primaryClientId,
+        }));
+      } else if (!primarySessionId || primarySessionId === 'default') {
+        set({ primaryClientId });
+      }
+      break;
+    }
+
+    case 'directory_listing': {
+      const cb = get()._directoryListingCallback;
+      if (cb) {
+        cb({
+          path: typeof msg.path === 'string' ? msg.path : null,
+          parentPath: typeof msg.parentPath === 'string' ? msg.parentPath : null,
+          entries: Array.isArray(msg.entries) ? msg.entries as DirectoryEntry[] : [],
+          error: typeof msg.error === 'string' ? msg.error : null,
+        });
+      }
+      break;
+    }
+
+    case 'file_listing': {
+      const fileBrowserCb = get()._fileBrowserCallback;
+      if (fileBrowserCb) {
+        fileBrowserCb({
+          path: typeof msg.path === 'string' ? msg.path : null,
+          parentPath: typeof msg.parentPath === 'string' ? msg.parentPath : null,
+          entries: Array.isArray(msg.entries) ? msg.entries as FileEntry[] : [],
+          error: typeof msg.error === 'string' ? msg.error : null,
+        });
+      }
+      break;
+    }
+
+    case 'file_content': {
+      const fileContentCb = get()._fileContentCallback;
+      if (fileContentCb) {
+        fileContentCb({
+          path: typeof msg.path === 'string' ? msg.path : null,
+          content: typeof msg.content === 'string' ? msg.content : null,
+          language: typeof msg.language === 'string' ? msg.language : null,
+          size: typeof msg.size === 'number' ? msg.size : null,
+          truncated: msg.truncated === true,
+          error: typeof msg.error === 'string' ? msg.error : null,
+        });
+      }
+      break;
+    }
+
+    case 'diff_result': {
+      const diffCb = get()._diffCallback;
+      if (diffCb) {
+        diffCb({
+          files: Array.isArray(msg.files) ? msg.files as DiffFile[] : [],
+          error: typeof msg.error === 'string' ? msg.error : null,
+        });
+      }
+      break;
+    }
+
+    case 'slash_commands': {
+      const slashSid = get().activeSessionId;
+      if (msg.sessionId && slashSid && msg.sessionId !== slashSid) break;
+      if (Array.isArray(msg.commands)) {
+        set({ slashCommands: msg.commands as SlashCommand[] });
+      }
+      break;
+    }
+
+    case 'agent_list': {
+      const agentSid = get().activeSessionId;
+      if (msg.sessionId && agentSid && msg.sessionId !== agentSid) break;
+      if (Array.isArray(msg.agents)) {
+        set({ customAgents: msg.agents as CustomAgent[] });
+      }
+      break;
+    }
+
+    case 'mcp_servers': {
+      const mcpTargetId = (msg.sessionId as string) || get().activeSessionId;
+      const servers = (msg.servers as McpServer[]) || [];
+      if (mcpTargetId && get().sessionStates[mcpTargetId]) {
+        updateSession(mcpTargetId, () => ({ mcpServers: servers }));
+      }
+      break;
+    }
+
+    case 'server_error': {
+      const allowedCategories = new Set<ServerError['category']>([
+        'tunnel', 'session', 'permission', 'general',
+      ]);
+      const category: ServerError['category'] =
+        typeof msg.category === 'string' && allowedCategories.has(msg.category as ServerError['category'])
+          ? (msg.category as ServerError['category'])
+          : 'general';
+      const message: string =
+        typeof msg.message === 'string' && (msg.message as string).trim().length > 0
+          ? stripAnsi(msg.message as string)
+          : 'Unknown server error';
+      const recoverable: boolean =
+        typeof msg.recoverable === 'boolean' ? msg.recoverable : true;
+
+      const serverError: ServerError = {
+        id: nextMessageId('err'),
+        category,
+        message,
+        recoverable,
+        timestamp: Date.now(),
+      };
+      set((state: ConnectionState) => ({
+        serverErrors: [...state.serverErrors, serverError].slice(-10),
+      }));
+      const errorMsg: ChatMessage = {
+        id: nextMessageId('err'),
+        type: 'error',
+        content: serverError.message,
+        timestamp: Date.now(),
+      };
+      const activeErrId = get().activeSessionId;
+      if (activeErrId && get().sessionStates[activeErrId]) {
+        updateActiveSession((ss) => ({
+          messages: filterThinking([...ss.messages, errorMsg]),
+          streamingMessageId: null,
+        }));
+      } else {
+        set({ streamingMessageId: null });
+        get().addMessage(errorMsg);
+      }
+      if (!serverError.recoverable) {
+        Alert.alert('Server Error', serverError.message);
+      }
+      break;
+    }
+  }
+}
+
+/** @internal Exposed for testing only — same pattern as _testQueueInternals */
+export const _testMessageHandler = {
+  handle: handleMessage,
+  setContext: (ctx: ConnectionContext) => { _connectionContext = ctx; },
+  clearContext: () => { _connectionContext = null; },
+};
