+++ packages/app/src/store/types.ts
@@ -1,0 +1,418 @@
+/**
+ * Shared type definitions for the connection store.
+ *
+ * Extracted from connection.ts to reduce file size and allow
+ * other modules (message-handler, utils) to import types without
+ * creating circular dependencies.
+ */
+
+/** Attachment metadata stored on a ChatMessage (base64 data cleared after send) */
+export interface MessageAttachment {
+  id: string;
+  type: 'image' | 'document';
+  uri: string;
+  name: string;
+  mediaType: string;
+  size: number;
+}
+
+export interface ChatMessage {
+  id: string;
+  type: 'response' | 'user_input' | 'tool_use' | 'thinking' | 'prompt' | 'error' | 'system';
+  content: string;
+  tool?: string;
+  options?: { label: string; value: string }[];
+  requestId?: string;
+  toolInput?: Record<string, unknown>;
+  toolUseId?: string;
+  toolResult?: string;
+  toolResultTruncated?: boolean;
+  /** Base64 images from tool results (e.g. computer use screenshots) */
+  toolResultImages?: { mediaType: string; data: string }[];
+  answered?: string;
+  expiresAt?: number;
+  timestamp: number;
+  /** Attachments on user_input messages (images, documents) */
+  attachments?: MessageAttachment[];
+  /** MCP server name (for tool_use messages from MCP tools) */
+  serverName?: string;
+}
+
+export interface SavedConnection {
+  url: string;
+  token: string;
+}
+
+export interface ContextUsage {
+  inputTokens: number;
+  outputTokens: number;
+  cacheCreation: number;
+  cacheRead: number;
+}
+
+export interface ClaudeStatus {
+  cost: number;
+  model: string;
+  messageCount: number;
+  contextTokens: string;
+  contextPercent: number;
+  compactPercent: number | null;
+}
+
+export interface InputSettings {
+  chatEnterToSend: boolean;
+  terminalEnterToSend: boolean;
+}
+
+export interface ModelInfo {
+  id: string;
+  label: string;
+  fullId: string;
+}
+
+export interface SessionInfo {
+  sessionId: string;
+  name: string;
+  cwd: string;
+  type: 'cli' | 'pty';
+  hasTerminal: boolean;
+  model: string | null;
+  permissionMode: string | null;
+  isBusy: boolean;
+  createdAt: number;
+  conversationId: string | null;
+}
+
+export interface DiscoveredSession {
+  sessionName: string;
+  cwd: string;
+  pid: number;
+}
+
+export interface DirectoryEntry {
+  name: string;
+  isDirectory: boolean;
+}
+
+export interface DirectoryListing {
+  path: string | null;
+  parentPath: string | null;
+  entries: DirectoryEntry[];
+  error: string | null;
+}
+
+export interface FileEntry {
+  name: string;
+  isDirectory: boolean;
+  size: number | null;
+}
+
+export interface FileListing {
+  path: string | null;
+  parentPath: string | null;
+  entries: FileEntry[];
+  error: string | null;
+}
+
+export interface FileContent {
+  path: string | null;
+  content: string | null;
+  language: string | null;
+  size: number | null;
+  truncated: boolean;
+  error: string | null;
+}
+
+export interface DiffHunkLine {
+  type: 'context' | 'addition' | 'deletion';
+  content: string;
+}
+
+export interface DiffHunk {
+  header: string;
+  lines: DiffHunkLine[];
+}
+
+export interface DiffFile {
+  path: string;
+  status: 'modified' | 'added' | 'deleted' | 'renamed' | 'untracked';
+  additions: number;
+  deletions: number;
+  hunks: DiffHunk[];
+}
+
+export interface DiffResult {
+  files: DiffFile[];
+  error: string | null;
+}
+
+export interface AgentInfo {
+  toolUseId: string;
+  description: string;
+  startedAt: number;
+}
+
+export interface ConnectedClient {
+  clientId: string;
+  deviceName: string | null;
+  deviceType: 'phone' | 'tablet' | 'desktop' | 'unknown';
+  platform: string;
+  isSelf: boolean;
+}
+
+export type SessionHealth = 'healthy' | 'crashed';
+
+export interface SessionContext {
+  gitBranch: string | null;
+  gitDirty: number;
+  gitAhead: number;
+  projectName: string | null;
+}
+
+export interface McpServer {
+  name: string;
+  status: string;
+}
+
+export interface SessionState {
+  messages: ChatMessage[];
+  streamingMessageId: string | null;
+  claudeReady: boolean;
+  activeModel: string | null;
+  permissionMode: string | null;
+  contextUsage: ContextUsage | null;
+  lastResultCost: number | null;
+  lastResultDuration: number | null;
+  isIdle: boolean;
+  health: SessionHealth;
+  activeAgents: AgentInfo[];
+  isPlanPending: boolean;
+  planAllowedPrompts: { tool: string; prompt: string }[];
+  primaryClientId: string | null;
+  conversationId: string | null;
+  sessionContext: SessionContext | null;
+  mcpServers: McpServer[];
+}
+
+export interface ServerError {
+  id: string;
+  category: 'tunnel' | 'session' | 'permission' | 'general';
+  message: string;
+  recoverable: boolean;
+  timestamp: number;
+}
+
+export interface SessionNotification {
+  id: string;
+  sessionId: string;
+  sessionName: string;
+  eventType: 'permission' | 'question' | 'completed' | 'error';
+  message: string;
+  timestamp: number;
+}
+
+export interface SlashCommand {
+  name: string;
+  description: string;
+  source: 'project' | 'user';
+}
+
+export interface CustomAgent {
+  name: string;
+  description: string;
+  source: 'project' | 'user';
+}
+
+export type ConnectionPhase =
+  | 'disconnected'        // Not connected, no auto-reconnect
+  | 'connecting'          // Initial connection attempt
+  | 'connected'           // WebSocket open + authenticated
+  | 'reconnecting'        // Auto-reconnecting after unexpected disconnect
+  | 'server_restarting';  // Health check returns { status: 'restarting' }
+
+/** Context captured from connect() closure for use by the extracted handleMessage(). */
+export interface ConnectionContext {
+  url: string;
+  token: string;
+  isReconnect: boolean;
+  silent: boolean;
+  socket: WebSocket;
+}
+
+/** Queued message for offline send buffer */
+export interface QueuedMessage {
+  type: string;
+  payload: unknown;
+  queuedAt: number;
+  maxAge: number;
+}
+
+export interface ConnectionState {
+  // Connection
+  connectionPhase: ConnectionPhase;
+  wsUrl: string | null;
+  apiToken: string | null;
+  socket: WebSocket | null;
+
+  // Saved connection for quick reconnect
+  savedConnection: SavedConnection | null;
+
+  // Server mode: 'cli' (headless) or 'terminal' (PTY/tmux)
+  serverMode: 'cli' | 'terminal' | null;
+
+  // Server context (from auth_ok)
+  sessionCwd: string | null;
+  serverVersion: string | null;
+  latestVersion: string | null;
+  serverCommit: string | null;
+
+  // Multi-session state
+  sessions: SessionInfo[];
+  activeSessionId: string | null;
+  sessionStates: Record<string, SessionState>;
+
+  // Legacy flat state (used when server doesn't send session_list, i.e. PTY mode)
+  claudeReady: boolean;
+  streamingMessageId: string | null;
+  activeModel: string | null;
+  permissionMode: string | null;
+  contextUsage: ContextUsage | null;
+  lastResultCost: number | null;
+  lastResultDuration: number | null;
+  isIdle: boolean;
+  messages: ChatMessage[];
+
+  // Available models from server (CLI mode)
+  availableModels: ModelInfo[];
+
+  // Available permission modes from server (CLI mode)
+  availablePermissionModes: { id: string; label: string }[];
+
+  // Discovered host tmux sessions (from discover_sessions)
+  discoveredSessions: DiscoveredSession[] | null;
+
+  // Claude Code status bar metadata (PTY mode)
+  claudeStatus: ClaudeStatus | null;
+
+  // Connected clients (multi-client awareness)
+  myClientId: string | null;
+  connectedClients: ConnectedClient[];
+  primaryClientId: string | null;
+
+  // Connection error feedback
+  connectionError: string | null;
+  connectionRetryCount: number;
+
+  // Server errors forwarded over WebSocket (last 10)
+  serverErrors: ServerError[];
+
+  // Background session notifications (permission, question, completed, error)
+  sessionNotifications: SessionNotification[];
+
+  // Shutdown state (reason + ETA for restarting banner countdown)
+  shutdownReason: 'restart' | 'shutdown' | null;
+  restartEtaMs: number | null;
+  restartingSince: number | null;
+
+  // Pending auto permission mode confirmation from server
+  pendingPermissionConfirm: { mode: string; warning: string } | null;
+
+  // Slash commands from server
+  slashCommands: SlashCommand[];
+
+  // Custom agents from server
+  customAgents: CustomAgent[];
+
+  // Directory listing callback for file browser
+  _directoryListingCallback: ((listing: DirectoryListing) => void) | null;
+
+  // File browser callbacks
+  _fileBrowserCallback: ((listing: FileListing) => void) | null;
+  _fileContentCallback: ((content: FileContent) => void) | null;
+
+  // Diff viewer callback
+  _diffCallback: ((result: DiffResult) => void) | null;
+
+  // View mode
+  viewMode: 'chat' | 'terminal' | 'files';
+
+  // Input settings
+  inputSettings: InputSettings;
+
+  // Raw terminal output buffer (ANSI-stripped, for plain text fallback)
+  terminalBuffer: string;
+
+  // Raw terminal buffer with ANSI codes intact (for xterm.js replay on view switch)
+  terminalRawBuffer: string;
+
+  // Imperative write callback for xterm.js (bypasses React state for performance)
+  _terminalWriteCallback: ((data: string) => void) | null;
+
+  // Actions
+  connect: (url: string, token: string, options?: { silent?: boolean; _retryCount?: number }) => void;
+  disconnect: () => void;
+  loadSavedConnection: () => Promise<void>;
+  clearSavedConnection: () => Promise<void>;
+  setViewMode: (mode: 'chat' | 'terminal' | 'files') => void;
+  addMessage: (message: ChatMessage) => void;
+  addUserMessage: (text: string, attachments?: MessageAttachment[]) => void;
+  appendTerminalData: (data: string) => void;
+  clearTerminalBuffer: () => void;
+  setTerminalWriteCallback: (cb: ((data: string) => void) | null) => void;
+  updateInputSettings: (settings: Partial<InputSettings>) => void;
+  sendInput: (input: string, wireAttachments?: { type: string; mediaType: string; data: string; name: string }[], options?: { isVoice?: boolean }) => 'sent' | 'queued' | false;
+  sendInterrupt: () => 'sent' | 'queued' | false;
+  sendPermissionResponse: (requestId: string, decision: string) => 'sent' | 'queued' | false;
+  sendUserQuestionResponse: (answer: string, toolUseId?: string) => 'sent' | 'queued' | false;
+  markPromptAnswered: (messageId: string, answer: string) => void;
+  markPromptAnsweredByRequestId: (requestId: string, answer: string) => void;
+  setModel: (model: string) => void;
+  setPermissionMode: (mode: string) => void;
+  confirmPermissionMode: (mode: string) => void;
+  cancelPermissionConfirm: () => void;
+  resize: (cols: number, rows: number) => void;
+
+  // Directory listing
+  setDirectoryListingCallback: (cb: ((listing: DirectoryListing) => void) | null) => void;
+  requestDirectoryListing: (path?: string) => void;
+
+  // File browser
+  setFileBrowserCallback: (cb: ((listing: FileListing) => void) | null) => void;
+  setFileContentCallback: (cb: ((content: FileContent) => void) | null) => void;
+  requestFileListing: (path?: string) => void;
+  requestFileContent: (path: string) => void;
+
+  // Diff viewer
+  setDiffCallback: (cb: ((result: DiffResult) => void) | null) => void;
+  requestDiff: (base?: string) => void;
+
+  // Session actions
+  switchSession: (sessionId: string) => void;
+  createSession: (name: string, cwd?: string) => void;
+  destroySession: (sessionId: string) => void;
+  renameSession: (sessionId: string, name: string) => void;
+  discoverSessions: () => void;
+  attachSession: (tmuxSession: string, name?: string) => void;
+  forgetSession: () => void;
+
+  // Slash commands
+  fetchSlashCommands: () => void;
+
+  // Custom agents
+  fetchCustomAgents: () => void;
+
+  // Full history sync (session portability)
+  requestFullHistory: (sessionId?: string) => void;
+
+  // Plan mode actions
+  clearPlanState: () => void;
+
+  // Server error actions
+  dismissServerError: (id: string) => void;
+
+  // Session notification actions
+  dismissSessionNotification: (id: string) => void;
+
+  // Convenience accessor
+  getActiveSessionState: () => SessionState;
+}
